requirements.txt
flask
openai
python-dotenv
gunicorn
requests
flask-cors
